+++ backend/src/models/teams.rs
@@ -5,13 +5,13 @@ use uuid::Uuid;
 pub struct TeamWithPoints {
     pub team_id: Uuid,
     pub name: String,
-    pub points: f64,
+    pub points: f32,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct CreateTeamRequest {
     pub name: String,
-    pub points: f64,
+    pub points: f32,
     pub week: i32,
     pub year: i32,
 }
